:recycle: #2 Change function name to createS3Backend to avoid complicated processing
