+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from "./createBackendTf";
+export * from "./createS3Backend";
